feat/user_context: Refactor auth endpoints and moved types to dedicated file
