+++ src/main.ts
@@ -1,11 +1,16 @@
-import express from "express";
+import express, { Application } from "express";
 import cors from "cors";
 import { config } from "./config";
 import authRoutes from "./routes/auth";
 import connectDB from "./db";
 
-const app = express();
-app.use(cors());
+const app: Application = express();
+app.use(cors({
+    origin: "http://localhost:8081", // Specify the exact origin of your React Native app
+    credentials: true,  // Allow cookies & authentication headers
+    methods: "GET,POST,PUT,DELETE", // Allowed methods
+    allowedHeaders: "Content-Type,Authorization" // Allowed headers
+}));
 app.use(express.json());
 
 connectDB();
