+++ src/routes/auth.ts
@@ -1,24 +1,12 @@
-import express, {Request, Response} from "express";
+import express, {Application, Request, Response} from "express";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcryptjs";
 import { config } from "../config";
 import { authenticateToken } from "../middleware/authMiddleware";
 import User from "../models/User";
+import { AuthenticatedRequest, LoginResponse, ProtectedRequest } from "../types";
 
 const router = express.Router();
-
-interface User {
-    email: string;
-    password: string;
-}
-
-const users: User[] = []; // Simulated Database
-
-interface RegisterRespnse {
-    message: string;
-    // status: (code: number);
-    // json: string;
-}
 
 // Register Route
 router.post("/register", async (req: Request, res: Response): Promise<void> => {
@@ -37,13 +25,7 @@ router.post("/register", async (req: Request, res: Response): Promise<void> => {
     const newUser = new User({ fullName, email, username, password: hashedPassword });
     await newUser.save();
     res.json({ message: "User registered successfully" });
-    return;
 });
-
-interface LoginResponse {
-    token?: string;
-    message?: string;
-}
 
 // Login Route
 router.post("/login", async (req: Request, res: Response<LoginResponse>) => {
@@ -52,24 +34,16 @@ router.post("/login", async (req: Request, res: Response<LoginResponse>) => {
     console.log("user found: " + user)
     if (!user || !(await bcrypt.compare(password, user.password))) {
         res.status(401).json({ message: "Invalid credentials" });
-
     }
     else {
         const token = jwt.sign({ email: user.email }, config.jwtSecret, { expiresIn: "1h" });
-        res.json({ token });
+        res.status(200).json({ token });
     }
 });
 
-interface ProtectedRequest {
-    user: User
-}
-
-export interface AuthenticatedRequest extends Request {
-  user?: ProtectedRequest; // Replace ProtectedRequest with the actual user type
-}
-
 // Protected Route
 router.get("/protected", authenticateToken, (req: AuthenticatedRequest, res: Response<ProtectedRequest>) => {
+    console.log("arrived");
     res.json(req.user);
 });
